requirements.txt
python-pptx
python-docx
epub2txt
gtts
